Improve pairings by taking into account the team score
Without such modification, the pairing algorithm had a tendancy to match the team
with the best score in a pool to the pool above. Good teams would have to face
more difficult adversaries which eventually leads to not so great final ranking
for them. The new algorithm produces a better final ranking by simply taking
into account the starting score within the same pool.